Handle the Black style for long argument lists.
Basically we just ignore `):` now.  This is something of a band-aid, but
I think it will always work for the Black formatter.  There are still
definitely ways to mess up the folding by being intentionally disruptive
(e.g. multiline argument list with dedented arguments, multiline data
structures with dedented entries), but this should handle the reasonable
cases.

To be more general, we'd like to to ignore anything within parentheses
(or braces or brackets) (over multiple lines) similarly to how multiline
strings are handled.  This is complicated by the fact that parentheses
can best nested, but it wouldn't be too hard to count the number of
parentheses that open or close on each line, and to ignore everything
until they match.

My concern with that is that I leave unmatched parentheses around when
developing code, and it would be really frustrating for that to ruin all
my folds, especially if I were editing in the middle of a big script.  I
don't think that's worth handling a borderline-malicious case more
accurately.

Fixes #4.