Debug vim8 native plugins.
I was having trouble getting the plugin to work with the vim8 native
plugin system, but that turned out to be because I was forgetting to set
the `filetype plugin indent on` option, not because of anything wrong
with plugin itself.  I added a note to the documentation to hopefully
clear this up for other people (and myself) in the future.