Allow trailing comment on lone arg list end line
When handling the case where a long argument list ends on its own line
at the same indentation level as 'def', allow a trailing comment on that
line. This commit adds \(#.\+\)\? to the end of the regex that handles
this case. The \+ is required so that this change doesn't interfere with
s:manual_open_pattern, where a lone trailing # will prevent folding on
that line.

def this_will_fold_properly_now(
    arg1,
    arg2,
    arg3,
): # some comment
    pass

def this_will_still_prevent_folding(
    arg1,
    arg2,
    arg3,
): #
    pass