Track parentheses and allow nested data structures.
Parentheses can affect how folds are made (e.g. #8, #29), tracking them
correctly is a challenge because parentheses that occur in strings have
to be ignored (and strings can be hard to parse).  This commit refactors
the line-parsing code make it possible to robustly track parentheses,
then implements the following improvements:

- Remove some old hacks for dealing with the parentheses in limited
  circumstances (e.g. black formatter corner case, import corner case).

- Add support for nested data structures.  Previously these folds
  couldn't be nested because a data structure was defined as an
  assignment, and there's no legal way to nest one assignment within
  another.  Now a data structure is just defined as a pair of
  parentheses/brackets/braces, so nesting is trivial.

- Fix #29