Wait for BufEnter before initializing.
CoiledSnake was being improperly initialized in diff-mode due to the
order in which autocommands were executed.  Specifically, the following
events (apparently) occur in the order given when loading a buffer:

- FileType autocommand
- &diff set
- BufEnter autocommand

Previously the plugin was initialized in response to FileType events,
but this was causing the wrong value of the &diff setting to be used.
Now, FileType events are only handled in BufEnter has already occurred.

Fixes #27