Ignore data structures that don't increase `paren_level`.
Data structures are detected by looking for changes in `paren_level`.
Previously, it was assumed that the first line of a data structure would
always have a lower `paren_level` than the last line, but it turns out
to be possible to consturct data structures where this is not the case.
For example:

  Line #  Paren Level  Code
  ======  ===========  =================
  1       0            x = [
  2       1            ] + [
  3       1            ]

The fold starting on line 1 ends on line 3, and has the expected
increase in `paren_level`.  However, a data structure fold also starts
on line 2 and does not have a line with `paren_level = 2` to end on.  As
a result, the fold extends to the end of the file and messes things up.

In some sense, coiled-snake is getting confused because it just adds up
all the parentheses on each line and is therefore not smart enough to
realize that there are two lists here.  But it another sense, treating
this syntax as a single data structure is probably the right thing to
do.

This commit tweaks the logic to end the fold on the same line it start
on if no increase in `paren_level` is found.  This has the practical
effect of removing the fold, since 0-length folds are not shown.